# api/openapi.yaml
openapi: 3.1.0
info:
  title: Alchemist API
  version: "1.0"
  description: |
    Video transcoding and delivery.

    Authenticate control-plane calls with `Authorization: Bearer <api-key>`.
    Playback URLs carry their own short-lived signature instead and need no key,
    because the caller is a viewer.

    Errors always take the shape `{"error":{"code","message"}}`. Branch on `code`,
    which is stable; `message` is written for a human and may be reworded or
    translated.
servers:
  - url: https://api.alchemist.example
paths:
...
  /admin/tenants:
    get:
      summary: List tenants
      description: Operator surface. Requires ALCHEMIST_ADMIN_KEY, not a customer key.
      responses:
        "200": { description: Tenants }
        "401": { description: invalid_admin_key }
        "404": { description: Admin surface disabled (no admin key configured) }
    post:
      summary: Create a tenant
      description: |
        Operator surface, behind a separate credential so a leaked customer key cannot
        mint tenants or keys.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                ladder_profile:
                  type: string
                  default: bd-mobile
                  description: Must be an existing profile.
      responses:
        "201": { description: Tenant created }
        "400": { description: invalid_request or invalid_ladder_profile }
        "401": { description: invalid_admin_key }
  /admin/contact:
    get:
      summary: Read contact enquiries
      description: Operator surface. Requires ALCHEMIST_ADMIN_KEY.
      responses:
        "200": { description: Enquiries, newest first }
        "401": { description: Missing or invalid admin key }
  /admin/tenants/{id}/keys:
    post:
      summary: Issue an API key for a tenant
      description: |
        The plaintext key is returned once and never again; only its hash is stored.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, default: default }
      responses:
        "201": { description: Key issued; api_key shown once }
        "404": { description: tenant_not_found }
  /admin/keys/{keyID}:
    delete:
      summary: Revoke an API key
      description: Takes effect on the next request; there is no cached credential.
      parameters:
        - { name: keyID, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Revoked }
        "404": { description: key_not_found }
  /v1/auth/signup:
    post:
      summary: Create an organisation and its first account
      security: []
      description: |
        Self-service. Creates the tenant, the owner, and one API key in a single
        transaction, and starts a session. The API key is in the response and is
        never shown again — only its hash is stored.

        Mounted only when ALCHEMIST_WEB_ORIGINS names at least one browser origin.
      responses:
        "201": { description: Account created, session started, API key returned }
        "400": { description: Missing org, malformed email, or password under 10 characters }
        "409": { description: An account already exists for that email }
        "429": { description: Too many attempts from this address }
  /v1/contact:
    post:
      summary: Send an enquiry from the public site
      security: []
      description: |
        No account needed — that is the point. Same origin check and rate limit as
        signup. A filled `website` field is treated as a bot and answered as though
        it worked.

        Requests are readable only through the operator surface: the table has RLS
        forced and no policy, so a tenant-scoped connection sees nothing.
      responses:
        "202": { description: Received }
        "400": { description: Missing name, bad email, or a message under 10 characters }
        "429": { description: Too many attempts from this address }
  /v1/auth/login:
    post:
      summary: Start a session
      security: []
      description: |
        A wrong password and an unknown address answer identically, and take the
        same time, so the endpoint cannot be used to find out who has an account.
      responses:
        "200": { description: Session started }
        "401": { description: Email and password do not match }
        "429": { description: Too many attempts from this address }
...
  /v1/auth/session:
    get:
      summary: Who the session belongs to
      security: []
      responses:
        "200": { description: User, tenant, email and organisation name }
        "401": { description: No live session }
  /v1/whoami:
    get:
      summary: Identify the authenticated tenant
      responses:
        "200": { description: Tenant identity and ladder profile }
        "401": { description: Missing or invalid API key }
  /v1/uploads:
    post:
      summary: Start a direct upload
      description: |
        Returns a presigned target. Upload the file there with `PUT`, then call
        `/v1/assets/{id}/complete`. Bytes never pass through this API.
      responses:
        "201": { description: Upload target created }
        "429": { description: Concurrency or monthly quota reached }
  /v1/assets:
    get:
      summary: Recent assets for the tenant
      description: Newest first. `limit` up to 100, default 25.
      responses:
        "200": { description: Assets with state, duration and size }
        "401": { description: Missing or invalid credentials }
    post:
      summary: Import a video from a URL
      description: |
        For migrating a library from another provider. The URL is re-validated at
        fetch time and on every redirect; addresses inside private, loopback or
        link-local ranges are refused.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string, format: uri }
      responses:
        "202": { description: Import queued }
        "400": { description: missing_url or invalid_url }
        "429": { description: Quota reached }
  /v1/assets/{id}:
    get:
      summary: Fetch an asset
      description: |
        `playback` appears once the state is `ready` or `partially_ready`. URLs are
        signed and expire.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Asset state and playback URLs }
        "404": { description: asset_not_found }
  /v1/assets/{id}/complete:
    post:
      summary: Signal that an upload has finished
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "202": { description: Queued for processing }
        "404": { description: No upload awaiting that id }
  /v1/keys:
    get:
      summary: List this tenant's API keys
      description: |
        Names and dates only. Keys are stored hashed, so an existing key cannot be
        shown again — a lost key is replaced, not recovered.
      responses:
        "200": { description: Keys, newest first }
        "401": { description: Missing or invalid credentials }
    post:
      summary: Mint another API key
      responses:
        "201": { description: The new key, shown once }
        "400": { description: Name too long }
        "401": { description: Missing or invalid credentials }
  /v1/keys/{id}:
    delete:
      summary: Revoke an API key
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Revoked }
        "404": { description: No such key, or it was already revoked }
  /v1/assets/{id}/chunks:
    get:
      summary: The chunk map for one asset
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      description: |
        Every chunk of every rendition, with its time range and whether it exists
        yet. This is what progress is made of.
      responses:
        "200": { description: Chunks, largest rendition first }
        "401": { description: Missing or invalid credentials }
  /v1/assets/{id}/activity:
    get:
      summary: What the pipeline did to one asset
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      description: |
        Queue steps with state, attempt count and timings. Deliberately not the
        error text: a queue failure carries whatever ffmpeg or the storage client
        said, including paths and internal arguments. The actionable part is the
        stable `error_code` on the asset itself.
      responses:
        "200": { description: Steps in the order they were queued }
        "404": { description: No such asset for this tenant }
  /v1/webhooks:
    get:
      summary: List webhook endpoints
      responses:
        "200": { description: Endpoints }
    post:
      summary: Register a webhook endpoint
      description: |
        The signing secret is returned once, here, and never again. Deliveries carry
        `X-Alchemist-Signature: sha256=<hmac>` over the raw body.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string, format: uri, description: Must be https }
                events:
                  type: array
                  items:
                    type: string
                    enum: [asset.ready, asset.failed, rendition.ready]
      responses:
        "201": { description: Endpoint created; secret shown once }
        "400": { description: invalid_url or unknown_event }
  /v1/bucket-sources:
    get:
      summary: List connected buckets
      responses:
        "200": { description: Sources with sync status and import counts }
    post:
      summary: Connect a customer-owned bucket
      description: |
        Ingests every video appearing under `prefix`. The secret is wrapped before
        storage and never returned again. A first scan starts immediately; afterwards
        a reconcile re-lists every 15 minutes, which is what makes the sync reliable
        when S3 event notifications are dropped or unsupported.

        Only known video extensions are taken, so PDFs and images in the same bucket
        are ignored. Re-listing never re-imports: objects are tracked by key and etag.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint, bucket, access_key_id, secret_access_key]
              properties:
                endpoint: { type: string, format: uri }
                region: { type: string, default: us-east-1 }
                bucket: { type: string }
                prefix: { type: string, default: "" }
                access_key_id: { type: string }
                secret_access_key: { type: string, description: Read access is sufficient }
      responses:
        "201": { description: Connected; first scan queued }
        "400": { description: invalid_url or invalid_request }
  /v1/auth/invite:
    get:
      summary: Who an invite is for
      description: |
        Unauthenticated — the point is to show the address and organisation before
        there is an account. An unknown token and an expired one give the same 404,
        so this cannot be used to probe which addresses were invited where.
      parameters:
        - { name: token, in: query, required: true, schema: { type: string } }
      responses:
        "200": { description: Email, role and organisation name }
        "404": { description: invite_not_found }
    post:
      summary: Redeem an invite and sign in
      description: |
        Creates the user inside the inviting tenant and starts a session. This is the
        only way a second user is ever created; signup always makes a new tenant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, minLength: 10 }
      responses:
        "201": { description: Account created and signed in }
        "400": { description: invalid_request or weak_password }
        "404": { description: invite_not_found }
        "409": { description: email_taken }
  /v1/members:
    get:
      summary: List teammates and pending invites
      description: |
        Session only. An API key is a machine credential and gets 403 `session_required`
        on every endpoint under `/v1/members`.
      responses:
        "200": { description: Members and pending invites }
        "403": { description: session_required }
  /v1/members/invites:
    post:
      summary: Invite someone to the account
      description: |
        The invite token is returned once, here, and never again — there is no mail
        sender in this repo yet, so the dashboard shows the link to pass on. Owner or
        admin only; only an owner can invite another owner. Invites last seven days,
        and a second invite to the same address replaces the first.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
                role: { type: string, enum: [owner, admin, member], default: member }
      responses:
        "201": { description: Invited; token shown once }
        "400": { description: invalid_email or invalid_role }
        "403": { description: session_required or not_permitted }
        "409": { description: already_a_member }
  /v1/members/invites/{id}:
    delete:
      summary: Withdraw a pending invite
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Withdrawn }
        "403": { description: session_required or not_permitted }
        "404": { description: not_found }
  /v1/members/{id}:
    patch:
      summary: Change a teammate's role
      description: |
        Only an owner can grant ownership, and the last owner cannot be demoted —
        an account with no owner is one nothing in the API can repair.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [owner, admin, member] }
      responses:
        "200": { description: Role changed }
        "400": { description: invalid_role }
        "403": { description: session_required or not_permitted }
        "404": { description: not_found }
        "409": { description: last_owner }
    delete:
      summary: Remove a teammate
      description: |
        Sessions cascade with the user, so this signs them out everywhere. You cannot
        remove yourself, and you cannot remove the last owner.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Removed }
        "403": { description: session_required or not_permitted }
        "404": { description: not_found }
        "409": { description: last_owner or cannot_remove_self }
...
  /v1/migrations:
    get:
      summary: Migrations you have started, and how far each has got
      description: |
        `total` is how many videos we found on the far side, `handled` how many of
        them we have finished with, so progress reads as a proportion rather than a
        bare number. `preview_done` says the list is complete and waiting on you.
      responses:
        "200": { description: Sources with total, handled, imported and skipped counts }
    post:
      summary: Migrate a library in from another video host
      description: |
        Session only — handing over another service's API key is account
        administration, not something a machine credential should be able to do.

        The key is encrypted before storage, bound to this one migration, and never
        returned. Nothing is imported yet: the migration starts in `previewing`,
        where we only page through the far side's library and write down what is
        there. Call `/confirm` once you have read the list and importing begins.

        Re-running is safe: videos are tracked by their remote id, so nothing is
        imported twice.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, secret]
              properties:
                provider: { type: string, enum: [vimeo, bunny, csv] }
                secret:
                  type: string
                  description: |
                    The provider's API key, or for `csv` the list itself — one link
                    per line, or a CSV with a url column and an optional title.
                config:
                  type: object
                  description: Non-secret settings the provider needs, such as a Bunny library id.
                  additionalProperties: { type: string }
      responses:
        "201": { description: Created in previewing; nothing is imported yet }
        "400": { description: unknown_provider or invalid_request }
        "403": { description: session_required }
  /v1/migrations/{id}:
    delete:
      summary: Forget a migration
      description: |
        Stops it and removes the source and its stored key. Videos already brought
        across are your assets now and are kept; nothing on the far side is touched.
        Session only.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Removed }
        "403": { description: session_required }
        "404": { description: not_found }
  /v1/migrations/{id}/confirm:
    post:
      summary: Approve the preview and start importing
      description: |
        Until this is called a migration has imported nothing — it has only listed
        what is there. This moves it to `scanning`, which is when assets are created
        and encoding is queued. Only valid once the preview has finished listing.
        Session only.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Now scanning }
        "403": { description: session_required }
        "404": { description: not_found }
        "409": { description: invalid_state }
  /v1/migrations/{id}/pause:
    post:
      summary: Stop taking new pages
      description: |
        The migration stops after the page it is on. Nothing already imported is
        undone. Session only.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Now paused }
        "403": { description: session_required }
        "404": { description: not_found }
        "409": { description: invalid_state }
  /v1/migrations/{id}/resume:
    post:
      summary: Carry on from where it stopped
      description: |
        Picks up at the stored cursor. Videos already across are not fetched again —
        they are matched on their remote id, and on their source URL account-wide.
        Session only.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Scanning again }
        "403": { description: session_required }
        "404": { description: not_found }
        "409": { description: invalid_state }
  /v1/migrations/{id}/items:
    get:
      summary: Which videos came across, and which did not
      description: |
        A video the far side will not hand over is recorded as skipped with a reason
        rather than failing the whole migration, so a partial result is explainable.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Up to 500 items, newest first }
...
  /v1/edits/{id}:
    delete:
      summary: Forget an edit
      description: |
        Removes the record only. The video it produced is an asset like any other and
        is deleted through the asset, not by discarding the edit that made it.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Forgotten }
        "404": { description: not_found }
...
  /v1/ladder-profile:
    put:
      summary: Choose the preset new uploads are built against
      description: |
        Existing assets keep the profile they were encoded with. Re-encoding a whole
        library because a setting moved would be a surprise that costs real money.
        Session only, owner or admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [profile]
              properties:
                profile: { type: string }
      responses:
        "200": { description: Changed }
        "400": { description: unknown_profile }
        "403": { description: session_required or not_permitted }
  /v1/playback-settings:
    get:
      summary: How your videos are delivered
      responses:
        "200": { description: Whether playback encryption is on }
    put:
      summary: Turn playback encryption on or off
      description: |
        **Off by default, deliberately.** Without a licence server this is not DRM:
        the manifest declares `KEYFORMAT="identity"` and the key is served from the
        same signed URL as the segments, so anyone able to fetch a segment can fetch
        the key. It adds nothing the expiring signed URL does not already give, and
        it makes playback impossible in every browser but Safari — Chrome and Firefox
        need EME, which needs a licence server.

        Access control is the signed expiring URL. Turn this on only if you have a
        reason to accept Safari-only playback.

        Videos already published keep whatever they were made with: re-packaging a
        library on a settings change would rewrite every object and evict the lot
        from every edge cache.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [encrypt_playback]
              properties:
                encrypt_playback: { type: boolean }
      responses:
        "200": { description: Changed; applies to the next video processed }
        "403": { description: session_required or not_permitted }
...
  /v1/branding/logo:
    put:
      summary: Replace the account logo
      description: |
        Raw image bytes, at most 1 MB, `Content-Type` one of image/png, image/jpeg or
        image/webp. The magic bytes must match the declared type. SVG is refused: it
        is a document that can carry script, and a logo is not worth that. Session
        only, owner or admin.
      requestBody:
        required: true
        content:
          image/png: { schema: { type: string, format: binary } }
          image/jpeg: { schema: { type: string, format: binary } }
          image/webp: { schema: { type: string, format: binary } }
      responses:
        "200": { description: Stored; returns the public URL }
        "400": { description: invalid_image }
        "403": { description: session_required or not_permitted }
        "413": { description: image_too_large }
    delete:
      summary: Remove the account logo
      responses:
        "204": { description: Removed }
        "403": { description: session_required or not_permitted }
...
  /v1/usage:
    get:
      summary: Billable usage for a period
      parameters:
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
      responses:
        "200": { description: Aggregated usage lines }
        "400": { description: invalid_date or invalid_range }
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
security:
  - apiKey: []